index.html to support tiled data

+++ bin/parallel/bashHelpers/parallel_runtools.sh
@@ -497,27 +497,44 @@ cat chr*/*/F3_orangeGraphs_${samplename}_${CCversion}/${flashstatus}_REdig_repor
 | sed 's/.*\s//' | awk 'BEGIN{a=0}{a=a+$1}END{print a}' > TMP_allRds.txt
 cat chr*/*/F3_orangeGraphs_${samplename}_${CCversion}/${flashstatus}_REdig_report_${CCversion}.txt | grep '16 Non-duplicated reads' \
 | sed 's/.*\s//' | awk 'BEGIN{a=0}{a=a+$1}END{print a}' > TMP_nondupRds.txt
+
+if [ "${tiled}" -eq 0 ];then
 cat chr*/*/F3_orangeGraphs_${samplename}_${CCversion}/${flashstatus}_REdig_report_${CCversion}.txt | grep '26a Actual reported fragments' \
 | sed 's/.*\s//' | awk 'BEGIN{a=0}{a=a+$1}END{print a}' > TMP_repFragTotal.txt
 cat chr*/*/F3_orangeGraphs_${samplename}_${CCversion}/${flashstatus}_REdig_report_${CCversion}.txt | grep '26b Actual reported CIS fragments' \
 | sed 's/.*\s//' | awk 'BEGIN{a=0}{a=a+$1}END{print a}' > TMP_repFragCis.txt
 cat chr*/*/F3_orangeGraphs_${samplename}_${CCversion}/${flashstatus}_REdig_report_${CCversion}.txt | grep '26c Actual reported TRANS fragments' \
 | sed 's/.*\s//' | awk 'BEGIN{a=0}{a=a+$1}END{print a}' > TMP_repFragTrans.txt
-
+else
+echo
+fi
+    
+if [ "${tiled}" -eq 0 ];then
 echo -e "allRds\tnondupRds\trepFragTotal\trepFragCis\trepFragTrans" > ${flashstatus}_dupFiltStats.txt
+else
+echo -e "allRds\tnondupRds" > ${flashstatus}_dupFiltStats.txt    
+fi
 
+if [ "${tiled}" -eq 0 ];then
 paste TMP_allRds.txt TMP_nondupRds.txt TMP_repFragTotal.txt TMP_repFragCis.txt TMP_repFragTrans.txt >> ${flashstatus}_dupFiltStats.txt
 rm -f TMP_allRds.txt TMP_nondupRds.txt TMP_repFragTotal.txt TMP_repFragCis.txt TMP_repFragTrans.txt
+else
+paste TMP_allRds.txt TMP_nondupRds.txt >> ${flashstatus}_dupFiltStats.txt
+rm -f TMP_allRds.txt TMP_nondupRds.txt 
+fi
 
 echo 'Nondup reads %' > ${flashstatus}_percentages.txt
 tail -n 1 ${flashstatus}_dupFiltStats.txt | cut -f 1-2 | awk '{print ($2/$1)*100}' >>${flashstatus}_percentages.txt
 echo '' >> ${flashstatus}_percentages.txt
+
+if [ "${tiled}" -eq 0 ];then
 echo 'Total cisreps/allrepfrags  %' >>${flashstatus}_percentages.txt
 tail -n 1 ${flashstatus}_dupFiltStats.txt | cut -f 3-4 | awk '{print ($2/$1)*100}' >>${flashstatus}_percentages.txt
 echo '' >> ${flashstatus}_percentages.txt
 echo 'Average reporter fragment count per read (final count)' >>${flashstatus}_percentages.txt
 tail -n 1 ${flashstatus}_dupFiltStats.txt | cut -f 2-3 | awk '{print ($2/$1)}' >>${flashstatus}_percentages.txt
 echo '' >> ${flashstatus}_percentages.txt
+fi
 
 cat ${flashstatus}_percentages.txt ${flashstatus}_dupFiltStats.txt > ${flashstatus}_percentagesAndFinalCounts.txt
  
@@ -2196,13 +2213,18 @@ echo '<br/> mappedR, multifragR, hascapR, singlecapF, withinSonicSizeF' >> index
 echo '</p>' >> index.html
 echo '' >> index.html
 
+capSiteName="capture site"
+if [ "${tiled}" -eq 1 ]; then
+    capSiteName="tile"
+fi
+
 echo '<pre>' >> index.html
 echo 'Abbreviations :' >> index.html
 echo '' >> index.html
 echo '         mappedR = mapped reads' >> index.html
 echo '      multifragR = reads with more than 1 fragment (can potentially report interaction)' >> index.html
-echo '         hascapR = reads containing fragment(s) overlapping any of the capture sites (within +/- sonicationSize from RE cut sites)' >> index.html
-echo '      singlecapF = fragment count in reads which can be resolved to a single capture site (not reporting multiple different capture sites within same read)' >> index.html
+echo '         hascapR = reads containing fragment(s) overlapping any of the '${capSiteName}'s (within +/- sonicationSize from RE cut sites)' >> index.html
+echo '      singlecapF = fragment count in reads which can be resolved to a single '${capSiteName}' (not reporting multiple different '${capSiteName}'s within same read)' >> index.html
 echo 'withinSonicSizeF = fragments within +/- sonicationSize from RE cut sites (filters out mapping errors)' >> index.html
 echo '</pre>' >> index.html
 echo '' >> index.html
@@ -2240,6 +2262,8 @@ echo 'NONFLASHED ' >> index.html
 echo '</b>' >> index.html
 echo 'ORANGE reads continue, light-orange reads are duplicates (filtered at this stage)' >> index.html
 
+if [ "${tiled}" -ne 0 ]; then
+
   fcountIN=$(cat ${rainbowRunTOPDIR}/D_analyseOligoWise/FLASHED_percentagesAndFinalCounts.txt    | grep -v '^\s*$' | tail -n 1 | cut -f 4,5 | sed 's/\t/,/')
  nfcountIN=$(cat ${rainbowRunTOPDIR}/D_analyseOligoWise/NONFLASHED_percentagesAndFinalCounts.txt | grep -v '^\s*$' | tail -n 1 | cut -f 4,5 | sed 's/\t/,/')
 
@@ -2255,6 +2279,9 @@ echo '<b style="color:orange">' >> index.html
 echo 'NONFLASHED ' >> index.html
 echo '</b>' >> index.html
 echo '</br>DARK color : cis reporters. LIGHT color : trans reporters.' >> index.html
+
+fi
+
 echo '</br>(hover over to see the counts)' >> index.html
 
 # -----------------------
@@ -2265,6 +2292,8 @@ echo '<h3>CCanalyser runs (duplicate filtering, final counts) :</h3>' >> index.h
 echo '<pre>' >> index.html
 cat ${rainbowRunTOPDIR}/D_analyseOligoWise/COMBINED_meanStdMedian_overOligos.txt >> index.html
 echo '</pre>' >> index.html
+
+if [ "${tiled}" -ne 0 ]; then
 
 countIN=$(head -n 2 ${rainbowRunTOPDIR}/D_analyseOligoWise/COMBINED_meanStdMedian_overOligos.txt | tail -n 1 | sed 's/.*max of:\s*//' | tr '\t' ',')
 echo '<h4>Reported fragments (total), oligo-wise distribution</h4>' >> index.html
@@ -2299,8 +2328,15 @@ tail -n 1 ${rainbowRunTOPDIR}/D_analyseOligoWise/COMBINED_meanStdMedian_overOlig
 echo '</pre>' >> index.html
 echo '' >> index.html
 
+fi
+
 countIN=$(head -n 1 ${rainbowRunTOPDIR}/D_analyseOligoWise/COMBINED_meanStdMedian_overOligos.txt | sed 's/.*max of:\s*//' | tr '\t' ',')
-echo '<h4>Capture fragments in reported reads (intra-read duplicates not filtered), oligo-wise distribution</h4>' >> index.html
+
+if [ "${tiled}" -ne 0 ]; then
+echo '<h4>Capture fragments in reported reads (intra-read duplicates not filtered yet in these counts), distribution over all capture sites</h4>' >> index.html
+else
+echo '<h4>Within-tile fragments in reported reads (intra-read duplicates not filtered yet in these counts), distribution over all tiles</h4>' >> index.html    
+fi
 echo '(hover over to see the counts)' >> index.html
 echo '<p>' >> index.html
 echo '<span class="boxplotprecalculated">'${countIN}'</span> ' >> index.html
